Fix import and patching weirdness by moving database helper into own module
